requirements.txt
Flask==2.0.1
PyMuPDF==1.20.0
transformers==4.28.0
python-docx==0.8.11
Pillow==9.2.0
pytesseract==0.3.10
keybert==0.7.0
langdetect==1.0.9
deep-translator==1.9.0
gunicorn==20.1.0
gevent==21.12.0